WIP minor bug fixes for build
