+++ AutoGladiators.Client/Services/PlayerMovementService.cs
@@ -1,6 +1,6 @@
 
 using System;
-using System.Timers;
+using System.Threading.Timer;
 using AutoGladiators.Client.Models;
 
 namespace AutoGladiators.Client.Services
